+++ CHANGELOG.md
@@ -49,3 +49,5 @@
 ### 2025-09-09--0917AM
 - 12MB configurable limit enforced (client + server)
 
+### 2025-09-09--1141AM
+- 12MB configurable limit enforced (client + server)